+++ NEWS.md
@@ -1,6 +1,11 @@
 # glmnettools 0.0
 
+## Changes in 0.0.2
+
+- Add `predict.rcv.glmnet`.
+- Add manual page for `plot.rcv.glmnet`.
+
 ## Changes in 0.0.1
 
 - Initial version.
-- Add `rcv.glmnet`.
+- Add `rcv.glmnet` and `plot.rcv.glmnet`.